requirements/requirements.dev.txt
black==20.8b1
isort==5.6.4
pytest==6.1.2
pytest-flask==1.1.0
pytest-flask-sqlalchemy==1.0.2
factory-boy==3.1.0
pytest-factoryboy==2.0.3
coverage==5.3
pytest-cov==2.10.1
pytest-sugar==0.9.4
